feature(tree): 支持select node #258
