+++ packages/tree/src/props.ts
@@ -179,6 +179,16 @@ export const treeProps = {
    * 节点拖拽时可交换位置（开启拖拽可交换位置后将不支持改变层级）
    */
   dragSort: PropTypes.bool.def(false),
+
+  /**
+   * 节点是否可以选中
+   */
+  selectable: PropTypes.bool.def(true),
+
+  /**
+   * 默认选中的节点id，selectable为false时无效
+   */
+  defaultSelectedNode: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).def(null),
 };
 
 type AsyncOption = {
