+++ packages/tree/src/tree.less
@@ -3,7 +3,7 @@
   .@{bk-prefix}-node-row {
     padding-left: calc(var(--offset-left));
 
-    &.is-checked {
+    &.is-selected {
       background-color: #ebf2ff;
     }
   }
